Clock fixes

+++ src/engines/misc/master/screen.hpp
@@ -9,7 +9,7 @@ namespace otto::engines::master {
   struct Screen : ui::Screen {
     void draw(nvg::Canvas& ctx) override;
     void action(itc::prop_change<&Props::volume>, float v) noexcept;
-    void action(itc::prop_change<&Props::tempo>, int t) noexcept;
+    void action(itc::prop_change<&Props::tempo>, float t) noexcept;
 
     float volume_ = 0.f;
     float rotation = 0.f;
